requirements/development.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==4.2.7
django-autoslug==1.9.9
django-dirtyfields==1.9.2
django-phonenumber-field==7.3.0
django-versatileimagefield==3.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.31
pillow==10.2.0
platformdirs==4.2.0
python-magic==0.4.27
pytz==2024.1
sqlparse==0.4.4
